requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
django-filter==25.1
python-decouple==3.8
psycopg2-binary==2.9.11
python-dotenv==1.1.1
gunicorn==23.0.0
mysqlclient==2.2.7
Markdown==3.9
django-cors-headers==4.9.0


# Postgres driver (needed if you use Postgres DATABASE_URL)
psycopg2-binary>=2.9
dj-database-url==3.0.1



setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
asgiref==3.9.2
certifi==2025.8.3
distlib==0.4.0
packaging==25.0
pipenv==2025.0.4
platformdirs==4.4.0